import p5-Test-Identity 0.01 OK benoit@
Comment:
assert the referential identity of a reference

Description:
This module provides a single testing function, "identical". It asserts
that a given reference is as expected; that is, it either refers to the
same object or is "undef". It is similar to "Test::More::is" except that
it uses "refaddr", ensuring that it behaves correctly even if the
references under test are objects that overload stringification or
numification.