import p5-Convert-Color-XTerm 0.06 OK benoit@
Comment:
indexed colors used by XTerm

Description:
This subclass of Convert::Color::RGB8 provides lookup of the colors
that xterm uses by default. Note that the module is not intelligent
enough to actually parse the XTerm configuration on a machine, nor to
query a running terminal for its actual colors. It simply implements
the colors that are present as defaults in the XTerm source code.