Remove security/nbaudit
No upstream, just a mirror left, no update since import in 1998, noone
else packages it.

We have net/nbtscan as a more recent alternative that is packaged by all
the major OSes as well;  it has a dead homepage and lacks behind in
updates, but still seems in better shape.

OK sthen