import ports/sysutils/py-subprocess-tee, ok solene@
This package provides a drop-in alternative to subprocess.run that
captures the output while still printing it in real-time, just the way
tee does.

Printing output in real-time while still capturing is valuable for any
tool that executes long-running child processes. For those, you do want
to provide instant feedback (progress) related to what is happening.