import ports/textproc/py-inflect, ok bket
inflect.py provides plural inflections, singular noun inflections,
"a"/"an" selection for English words, and manipulation of numbers as
words.

Plural forms of all nouns, most verbs, and some adjectives are provided.
Where appropriate, "classical" variants (for example: "brother" ->
"brethren", "dogma" -> "dogmata", etc.) are also provided.

Single forms of nouns are also provided. The gender of singular pronouns
can be chosen (for example "they" -> "it" or "she" or "he" or "they").

Pronunciation-based "a"/"an" selection is provided for all English
words, and most initialisms.

It is also possible to inflect numerals (1,2,3) to ordinals (1st, 2nd,
3rd) and to English words ("one", "two", "three").

In generating these inflections, inflect.py follows the Oxford English
Dictionary and the guidelines in Fowler's Modern English Usage,
preferring the former where the two disagree.