Update to UEFITool A60
Upstream's versioning is a bit confusing;  there has been no 0.x.y release in
two years, but they keep shipping A* versions, so jump from 0.28.0 to 60.

This now only includes the graphical UEFITool built with qmake.
The CLI tools are built with cmake and could be readded later on, but for now
fiddling with both qmake and cmake modules in one port is too much hassle.

A60 is the latest version that compiles without errors/patches.