Import devel/crc32c
Initially ported to avoid the bundled copy in net/tg_owt, but other ports
can probably pick it up as well.

Feedback OK sthen tb
---
Collection of CRC32C implementations that dispatches to one suitable to
the hardware. CRC32C is the cyclic redundancy check that uses the iSCSI
polynomial specified in RFC 3720.