update to fnaify 3.1 Release Notes: https://github.com/rfht/fnaify/releases/tag/3.1
Main change is that it no longer removes bundled Steamworks.NET.dll. If
you want to run a game that usually bundles it, but has been installed
or run with a previous version of fnaify, you may need to do this:

$ mv fnaify-backup/Steamworks.NET.dll .

Tested with most of the catalogue of FNA/fnaify supported games.