Pull two upstream fixes
--ignore-permission-errors didn't work in all cases, so I fixed one.

Testing my PR on top of 0.12.0 as well and not just master revealed another
bug, which lead to truncated filenames.

This is fixed in master already.

Do not wait for another release and apply both now, so backups no longer
abort or corrupt in those use cases.

All 57 (58 with this diff) bats(1) tests pass on amd64.

OK bket