update findlib to 1.9.6; needed for ocaml > 4.12
ocaml 4.13 changed capitalization of some of the lib files which confuses
some of the findlib build tests (and then leads to fallback code which
depends on ocamlbuild).

A newer version of findlib addresses this.

In addition, findlib 1.9.4 added proper DESTDIR support, so we can drop
FAKE_FLAGS from the port Makefile as part of this update.

Diff from Volker Schlecht; and tested by myself on amd64.