urxvt OSC colour queries were not terminated correctly; they should use ST or BEL but only actually used a single ESC. Hack to replace with BEL for now while I report upstream.
The combination of the recent update to this port (which fixed a bug
that prevented urxvt from responding to colour queries other than OSC 4)
and a recent change in tmux (which started querying fg/bg colour with
OSC 10/11 respectively) resulted in this being particularly annoying.

Cause diagnosed by nicm, thanks!