Drop autoconf patch, use saner cmake
cmake gets the PREFIX stuff right and is easier to work with.

Qt5Svg is always REQUIRED in the global cmake file but only used in src/app/
for Windows and macOS, so make it a build dep.

With cmake the dbus support bits are under 'if (LINUX)', so enable dbus
explicitly to retain notification and power management support.

Both autoconf and cmake pick up <execinfo.h> and enable stacktrace printing
on crashes by default, but only autoconf manages to pass -libexecinfo;
I'll send a patch upstream in case future updates don't solve the cmake
bits for dbus and stacktrace.

maintainer timeout
OK rsadowski