Adjust build-debug-info (used to create debug packages):
- Use llvm-objcopy instead of (binutils) objcopy when creating detached
debug symbols and adding the gnu-debuglink section.

- For llvm-objcopy and strip, output to a temporary file and copy it
back to the original file. The LLVM tools create a new file, so hard links
are not normally maintained, but this method keeps them.

https://github.com/llvm/llvm-project/issues/59848#issuecomment-1374070557

ok naddy@.

This should allow us to move back to llvm-strip in base on archs where
we use ld.lld, which (alongside the change to llvm-objcopy) avoids problems
that have been seen during mimmutable(2) development with GNU_RELRO with
the version of binutils that we have.