x11/xpostit: explicitly cast integer to pointer to fix build with clang 15
The callback takes a pointer to be able to pass arbitrary data, but
instead of allocating an int and passing a pointer to it, the code
jams the int directly into the pointer for an int -> pointer -> int
roundtrip.  Problem reported by and discussed with jca@.

Fix implicit function declarations.