seamonkey: unbreak build by removing the fft.asm patch.
naddy reported a build breakage and noted that there is a difference
between the patch I applied to the mozillas and kettenis's patch. For
some reason this worked for the other mozilla variants, but not here.
Since seamonkey needs USE_NOEXECONLY due to skia, the next build will
work despite the missing asm fix.