Neuter -Wint-conversion to unbreak build with clang 15
One function in the vala source code ends up as a C function that
returns NULL instead of an int.  It's not obvious to me how to fix the
vala source in a semantically correct way, so just downgrade the error
to a warning.  Hopefully the generated code is still correct.
ok aja@ (maintainer)