update lang/guile3 to 3.0.9
announce:
https://www.gnu.org/software/guile/news/gnu-guile-309-released.html

The new `spawn' and existing functions like `system*', `open-pipe' and
`pipeline' are now implemented via posix_spawn, except that the way they
run on non-linux OSes is wrong.  The closing of _all_ file descriptors
in the 3...RLIMIT_NOFILE is scheduled via posix_spawn_file_actions_addclose
and, since not all the fds in that range are valid, close(2) fails and aborts
posix_spawn.  Add an ugly hack to close only in the 3...getdtablecount()
range, testing each fd with fstat(2) to see if it's 'live'.
Bug filed upstream, waiting for an official fix.