Import sysutils/rustic-0.4.3
OK rsadowski@, aisha@

DESCR
Rustic is a backup tool that provides fast, encrypted, deduplicated
backups. It reads and writes the restic repo format desribed in the
design document and can therefore be used as a complete replacement for
restic.

Note that rustic currently is in an beta release and misses tests. It is
not yet considered to be ready for use in a production environment.