p5-Sereal-* - Update description with warning about upgrading decoder first
The Sereal tooling is provided in two packages, one for decoding
and one for encoding so that the decoder may be safely upgraded
before and independently of the encoder. This allows the package
to be safely upgraded in multi-host contexts where it is impossible
to upgrade all machines simultaneously. Upgrading the encoder
before upgrading the decoder may result in data loss in such an
environment, so add a note saying as much.