rust-ring: add aarch64 xonly support
This ports changes landed in BoringSSL to add aarch64 xonly support to
ring. There has been some divergence between the files, so the diff is
somewhat different. Missing are the extra contortions needed in perlasm
to accomodate the special needs of the Apple ecosystem. At least this
now passes regress on OpenBSD.

https://boringssl-review.googlesource.com/c/boringssl/+/57446