update pysmt to 0.9.5 which added python 3.10 support
The current version of PySMT is broken at runtime due to use of imports
that were removed in Python 3.8. Fix this by updating to a newer version
of PySMT.

Add a note that six is no longer needed as an RDEP for PySMT, also add
PORTHOME to make the tests work slightly better.

Submitted by Todd Carson with tweaks by me; thanks.