Update to redis-6.2.11
This includes fixes for:

(CVE-2023-25155) Specially crafted SRANDMEMBER, ZRANDMEMBER, and
HRANDFIELD commands can trigger an integer overflow, resulting in a
runtime assertion and termination of the Redis server process.
(CVE-2022-36021) String matching commands (like SCAN or KEYS) with a
specially crafted pattern to trigger a denial-of-service attack on
Redis, causing it to hang and consume 100% CPU time.