multimedia/libv4l: __syscall(2) is gone. Use mmap(2) instead.
This fixes the build of graphics/ffmpeg on ld.bfd architectures.
This will have to be revisited once we endeavor to kill syscall(2),
but that's not going to happen right now.

It is quite sad that our old crappy ld.bfd is better in may ways
than the shiny new ld.lld: it seems much less prone to producing
artefacts with runtime breakage.

ok kn (who had a similar diff)