archivers/zstd: unbreak on strict alignment architectures
For compilers defining __GNUC__, the new zstd defaults to using versions
of uint16_t, uint32_t, and uint64_t with __attribute__((aligned(1))).
That doesn't end well on sparc64, for example. Disable this default
unaligned access everywhere and fall back to memcpy(). This unbreaks
cmake's zstd detection and with it x11/qt6/qtbase among other things.

Also fix a crash due to a setvbuf() call with a NULL FILE * that
happens for example if the user has insufficient permissions to
create the output file.

ok bket (maintainer) sthen