p5-Crypt-OpenSSL-DSA: don't set dummy public key to 0
When the Richards added setters for DSA parameters to OpenSSL 1.1, they
added a check that disallows setting a private key without a corresponding
public key. This was incompatible with existing API surface of the Perl
bindings, so they added a hack. This hack no longer works since we added
a check to the DSA signing operation.

Set the public key to 2 instead, which is good enough to pass the new
checks. This workaround will have to be revisited soon, but it is good
enough for now.

Discussed at length with beck, bluhm, jsing
reported by and ok bluhm (maintainer)