openssl: disable assembly on sparc64
On an LDOM on a T4-2 I see crashes in about 2 out of 3 regress test runs,
likely due to an unaligned access in the T4-specific perlasm. I only saw
crashes in AES-CBC but they might also affect other AES block cipher modes
as well as DES and Camellia. This is unacceptable for a dependency of a
mailserver and a backup tool where reliability is key.

Reported upstream in https://github.com/openssl/openssl/issues/20523

ok kn phessler