databases/influxdb: update to 1.11.0.
on top of the update patch go.{sum,mod} to force-update x/sys module to
0.5.0. With that influxdb binary starts fine and stops crashing with
SIGSYS all over the place.

Note that since the flux component is partly written in rust and in go,
this port now also uses devel/cargo module in addition to the go module,
making it a nice frankeinstein monster.