update to traccar-5.7, from Renaud Allard (maintainer)
Note that default.xml has changed a path; if you have modified it (which
is not recommended, but might have been done to disable unwanted protocols)
you need to merge the changes.

Update pkg-readme to show a better way to disable protocols that avoids
the need to touch default.xml (this part from me).