import games/brogue from Anton Konyahin (me () konyahin DOT xyz) - thanks! I added NO_TEST=Yes ok op@
DESCR:
Brogue is a single-player strategy game set in the halls of a
mysterious and randomly-generated dungeon. The objective is simple
enough - retrieve the fabled Amulet of Yendor from the 26th level -
but the dungeon is riddled with danger. Horrifying creatures and
devious, trap-ridden terrain await. Yet it is also riddled with
weapons, potions, and artifacts of forgotten power. Survival demands
strength and cunning in equal measure as you descend, making the most
of what the dungeon gives you. You will make sacrifices, narrow
escapes, and maybe even some friends along the way - but will you be
one of the lucky few to return alive?