Disable stack guard pages in the primordial thread only to fix java integration with libreoffice.
The primordial thread's stack permissions are immutable so stack
guard pages can not be placed on it. Fortunately the launcher for
normal java programs uses a separate thread to start the JVM and
stack guard pages may be placed on thread stacks. Libreoffice, on the
other hand, launches the JVM in the primordial thread and stack guard
pages fail to be placed. Disabling stack guard pages for the primordial
thread allows libreoffice to use java again, but with the side effect of
breaking StackOverflowError exceptions on the primordial thread only.