import victor-mono
Victor Mono is an open-source monospaced font with optional
semi-connected cursive italics and programming symbol ligatures.

The typeface is slender, crisp and narrow, with a large x-height
and clear punctuation, making it legible and ideal for code. It
comes in seven weights and Roman, Italic and Oblique styles.

package by Ashlen <dev AT anthes DOT is> who is the maintainer

ok op