update to icinga2-2.14.0, see notes at https://icinga.com/blog/2023/07/12/releasing-icinga-2-14-and-2-13-8/ (there are a few breaking changes, though many users won't need to do anything special)
port tweak: switch to the non-unity build on all archs (was only done
on !lp64 and for debug builds before) and set DPB_PROPERTIES=parallel.

big thanks to Alexander Klimov for fiddly work upstream in this release
cycle to have Icinga working better on OpenBSD (in particular related to
pid not being filled-in in our siginfo, and rc.d system use of process
strings rather than pid files, these have been carried in patches here
for a while).