openvpn: explicitly turn off ENGINE support
We will need to provide stubs for some projects that don't allow turning
off ENGINE (looking at you M2Crypto), and these stubs are enough for the
configure script of openvpn to deduce that ENGINE support is available.
Just turn this off so we don't need to add another half a dozen stubs.

ok jca