ldns: forcibly set HAVE_OPENSSL_ENGINE_H to no and explicitly disable GOST
GOST is currently disabled since the port is built with OPENSSL_NO_ENGINE.
With the next bump the strange logic of header guards in this code will end
up enabling GOST, which in turn will break the build. So unbreak the build
and disable GOST to keep things working.

ok sthen