update to isc-bind-9.18.17
- If a response from an authoritative server has its RCODE set to
FORMERR and contains an echoed EDNS COOKIE option that was present
in the query, named now retries sending the query to the same server
without an EDNS COOKIE option. [GL #4049]

- The relaxed QNAME minimization mode now uses NS records. This reduces
the number of queries named makes when resolving, as it allows the
non-existence of NS RRsets at non-referral nodes to be cached in
addition to the normally cached referrals. [GL #3325]

- The ability to read HMAC-MD5 key files, which was accidentally lost in
BIND 9.18.8, has been restored. [GL #3668] [GL #4154]

- Several minor stability issues with the catalog zone implementation
have been fixed. [GL #4132] [GL #4136] [GL #4171]