teach portroach to use the cpan handler for more cpan MASTER_SITES lines (https was not handled, and it required a /CPAN/modules/ path, so also allow https and {meta,}cpan.org/modules to match).
also, when stripping versions from distnames for the cpan handler to pass
to a metacpan search, cope with Name-v1.23 format as well as Name-1.23.

this avoids falling back to directory lists in some cases, which is
particularly important for cpan because their version number ordering
does not match the ordering normally used in ports/portroach, e.g.
1.1, 1.101, 1.2 vs 1.1, 1.2, 1.101, and also portroach's listing
handler doesn't cope with a mixture of -1.23 and -v1.23 formats in
the same port either.