Introduce MODCARGO_INSTALL_TARGET_PATH
For virtual manifests, 'cargo install --path .' doesn't work during the
'make fake' step. For this, one needs to be able to specify a path which
is typically the name of the executable one wants to package. Lets landry
update mail/meli, and will also help sthen with the leftwm wip.

fine with semarie
ok landry