Stop building the tree-sitter cli program to drop the dep on lang/rust
Upstream intends this library to be dep-free so that it can be used in
any project.  We want to be able to use tree-sitter in the editors/emacs
port, but rust is not portable enough and a too big dependency for
emacs--no_x11.  So just drop the tree-sitter program, which is separate
from the library and only useful to develop new grammars.

ok paco@ sthen@, "go ahead" edd@