import editors/novelwriter changes and ok sthen@ ok volker@
DESCR:
novelWriter is a plain text editor designed for writing novels assembled from
many smaller text documents. It uses a minimal formatting syntax inspired by
Markdown, and adds a meta data syntax for comments, synopsis, and
cross-referencing. It's designed to be a simple text editor that allows for easy
organisation of text and notes, using human readable text files as storage for
robustness.

The project storage is suitable for version control software, and also well
suited for file synchronisation tools. All text is saved as plain text files
with a meta data header, the core project structure is stored in a single
project XML file, and other meta data is primarily saved as JSON files.