Update nerd-fonts to 3.0.2, patch from Joel Carnat, thanks!
We're going from version 2.2.2 to 3.0.2.

Breaking Changes:

1. This release fixes some long standing issues that are due to the
naming of the fonts: There is a completely new naming scheme. This might
be inconvientient for existing setups.

2. The old Material Design Icon codepoints are finally dropped. Due
to an historic mistace we placed them in between some asiatic glyphs,
breaking that script Since v2.3.0 the (updated and expanded) Material
Design Icons have new codepoints in the 5 digit region.

More details are described in the 3.0.0 change log:
https://github.com/ryanoasis/nerd-fonts/releases/tag/v3.0.0

ok op@