Unbreak graphics/imlib2 for arch without rust
If there's no rust compiler, then x11/gnome/librsvg is an old version
and fails imlib2's check for librsvg-2.0 >= 2.46; the built imlib2 is
missing svg.so and was failing to package.  Unbreak with a conditional
comment in PLIST, so we can package imlib2 without svg.so if we don't
have rust, or with svg.so if we do.

Also add sharpyuv to WANTLIB.  I redid WANTLIB on a rust arch; the
trick is that "make RUST_ARCHS=" pretends to not have rust, so I can
twice run port-lib-depends-check (without then with rust).

help from semarie@
ok sdk@
older diff ok op@