import textproc/simpleini
SimpleINI is a single-header, cross-platform library that provides a
simple API to read and write INI-style configuration files. It supports
data files in ASCII, MBCS and Unicode. It is designed explicitly to be
portable to any platform and has been tested on Windows, WinCE and
Linux.

feedback from and ok sthen@