Remove ruby-kramdown
Broken in all supported Ruby versions due to a missing dependency on
rexml (which was removed from the Ruby standard library in Ruby 3).
Better to remove this than import and maintain a new port for rexml.

Pure Ruby port, and nothing depends on it.  If you need this, use gem
install.

OK kn@, tb@