Import ksshaskpass-5.27.7
Comment:
KDE Plasma ssh-add helper

Description:
Ksshaskpass is a front-end for ssh-add which stores the password of the sh key
in KWallet.

Ksshaskpass is not meant to be executed directly, you need to tell ssh-add
about it. ssh-add will then call it if it is not associated to a terminal.

Maintainer: Rafael Sadowski <rsadowski@openbsd.org>

OK kn@, tested by volker@