Update to Monkey's Audio Codec 10.20
This switches to using the upstream distfile, instead of what appears
to be a Linux port of a quite old version (Monkey's Audio was
previously Windows-only). The upstream distfile compiles without
hanges.

The shared library has changed from libmac to libMAC.  The license
has changed from a custom, non-open source license to a standard
BSD-3-Clause license.

All previous patches no longer necessary or apply. Add a couple
new patches to the include files to make it easier to use the
library.