Remove sysutils/borgbackup/1.1
The 1.1 branch is considered EOL by upstream. The 1.2 branch, which we
have in ports as well, is actively maintained, and upgrading from 1.1.x
to 1.2.x is possible as detailed by
https://github.com/borgbackup/borg/blob/1.2.6/docs/changes.rst#borg-11x-to-12x.

Added @pkgpath markers to the PLIST of sysutils/borgbackup/1.2 so that
pkg_add -u will update borgbackup-1.1.x to borgbackup-1.2.x. Bumped
REVISION of 1.2.

OK kn@