Forcibly disable hidden dep on Wayland.
The upstream knob to turn it off doesn't work, so Wayland is used
unconditionally if installed, necessitating patching the knob.

Even if it weren't junked during the build, making Wayland a dependency
of this port is not an option right now, as it requires SDL2 to have
been built with Wayland support.

Build breakage reported by ajacoutot@