Switch from self-hosted distfile to GitHub DIST_TUPLE.
The previous self-hosted distfile was mistakenly made from a newer
upstream commit than the latest release tag. That means that using the
release tag for DIST_TUPLE now would be incompatible with existing game
saves, so until a new release is made, we'll use that slightly newer
upstream commit for the OpenBSD package.

ok kirby@