import p5-Hash-Ordered 0.014 from Wen Heping; OK benoit@
Comment:
fast, pure-Perl ordered hash class

Description:
This module implements an ordered hash, meaning that it associates keys
with values like a Perl hash, but keeps the keys in a consistent order.
Because it is implemented as an object and manipulated with method
calls, it is much slower than a Perl hash. This is the cost of keeping
order.