haproxy: enable QUIC support
A substantial amount of effort was spent over a year ago to make QUIC
work in libssl. haproxy upstream has gone to quite some length to make
this work with LibreSSL and they run their CI with it. To make use it,
users still have to opt in using the quic{4,6}@: keywords and the h3 alpn:

https://www.haproxy.com/blog/how-to-enable-quic-load-balancing-on-haproxy

ok danj jsing