comms/hylafax: mark as BROKEN
upstream has had two commits in 2018 and one in 2014, otherwise the last
update was in 2012. It relied on the now-not-build-by-default tiff2ps
and tiff2pdf binaries.

ppl actually using hylafax on OpenBSD should speak up to improve that
situation..there's apparently a fork at https://hylafax.sourceforge.io/
but i have no idea if this has better support for modern tiff (i doubt so)

ok tb@