Import textproc/urlscan 1.0.1
Urlscan is a small program that is designed to integrate with the
"mutt" mailreader to allow you to easily launch a Web browser for URLs
contained in email messages. It is a replacement for the "urlview"
program with additional features.

Port created by me, with help from Laurent and hint from Stuard.

OK sthen, lraab