import new sysutils/eza
eza is a replacement for ls written in Rust. Its options are similar to the
regular ls, but it also adds lots of fancy features like tree traversing,
colors, git support, sorting options..

eza is a fork of exa (which is no longer maintained).

initial work from sthen@
ok to import sthen@, tb@