Retire OpenSSL 3.0, ok rsadowski sthen
If you don't care about the FIPS or LTS selling points of OpenSSL 3.0,
then there is no reason to use this. It also doesn't have landing pads
in its assembly, making it a risk on very modern hardware. Thankfully,
the bit bucket in the sky is used to all sorts of gruesome horrors, so
it won't mind this particular one.

OpenSSL 3.1 is going to live until pi day 2025, which gives us at least
two full releases worth of 3.1 (including stable) and until then there
is some chance that there will be 3.2 with the QUIC MVP.